table

+++ app/Http/Controllers/TasksController.php
@@ -15,10 +15,24 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all();
-        
-        return view('tasks.index',['tasks' => $tasks,]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index',$data);
+        } 
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -43,15 +57,15 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'user_id' =>'required|max:255',
+            
             'status'=>'required|max:10',
             'content' =>'required|max:255',]);
         
-        //
+      // メッセージを作成
         $task = new Task;
-        $task->status = $request->status;
-        $task->user_id = $request->user_id;
         $task->content = $request->content;
+        $task->user_id = \Auth::id();
+        $task->status = $request->status;
         $task->save();
 
         // トップページへリダイレクトさせる
@@ -67,10 +81,18 @@ class TasksController extends Controller
     public function show($id)
     {
         //
+       
         $task = Task::findOrFail($id);
-        
-        return view('tasks.show',['task'=>$task,]);
+         // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show',['task'=>$task,]);
+            
+        }
+
+        // 前のURLへリダイレクトさせる
+        return redirect('/');
     }
+    
 
     /**
      * Show the form for editing the specified resource.
@@ -82,7 +104,13 @@ class TasksController extends Controller
     {
         //
         $task = Task::findOrFail($id);
-        return view('tasks.edit',['task'=>$task,]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit',['task'=>$task,]);
+            
+        }
+
+        // 前のURLへリダイレクトさせる
+        return redirect('/');
     }
 
     /**
@@ -95,17 +123,20 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
-            'user_id' => 'required|max:255',
+            
             'status' => 'required|max:10',
             'content' => 'required|max:255',]);
         //
-         $task = Task::findOrFail($id);
+        $task = Task::findOrFail($id);
         $task ->status = $request->status;
-        $task ->user_id = $request->user_id;
         $task->content = $request->content;
-        $task->save();
+         if (\Auth::id() === $task->user_id) {
+            $task->save();
+            
+        }
 
-       
+        // 前のURLへリダイレクトさせる
+
         return redirect('/');
     }
 
@@ -117,12 +148,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        //
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
         
     }
